Reconsider my approach to importing names
Since the whole tool is about Git, it makes sense to import git2 names.
Yet, some of them (e.g. Error) are wa-ay too generic, so I kept them
non-imported.

HashMap is used throughout the program, so its name is imported to
simplify signatures.

The rest of the names I un-imported to make the code clearer.