Correct the picture in README
We merge submodule's history into main repo's, but never the other way
around. This helps preserve the "promise" that submodules given: the
work in submodule is completely detached from the work in the main repo.
If we were to inherit submodule's history from the main repo's, we'd be
making another "promise", a one saying "here's submodule's code *plus
main repo's code* that we know is okay".