Drop invalid TODO item
We can't use TreeBuilder in rewrite_submodule_history because
TreeBuilder doesn't implement Iterator trait, and we need it in order to
modify all paths. We could emulate this with TreeBuilder's filter() that
modifies second TreeBuilder, but this isn't any different from the
approach we already have. So we'll stick with what already works.

Closes #10.